CareSet/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ezdxf>=1.4.3",
    "numpy>=2.3.5",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "streamlit-drawable-canvas>=0.9.3",
    "streamlit-plotly-events>=0.0.6",
]
